requirements.txt
customtkinter>=5.0.0
matplotlib>=3.7.0
speedtest-cli>=2.1.3

